+++ packages/cli/src/info.ts
@@ -5,17 +5,11 @@
  */
 
 import { resolveLanguageModelConfigurations } from "../../core/src/config"
-import { parseTokenFromEnv } from "../../core/src/connection"
-import { MODEL_PROVIDERS } from "../../core/src/constants"
-import { errorMessage } from "../../core/src/error"
 import { host, runtimeHost } from "../../core/src/host"
-import { resolveLanguageModel } from "../../core/src/lm"
 import {
     ModelConnectionInfo,
     resolveModelConnectionInfo,
 } from "../../core/src/models"
-import { LanguageModelConfiguration } from "../../core/src/server/messages"
-import { deleteEmptyValues } from "../../core/src/util"
 import { CORE_VERSION } from "../../core/src/version"
 import { YAMLStringify } from "../../core/src/yaml"
 import { buildProject } from "./build"
