+++ packages/core/src/llms.ts
@@ -2,7 +2,7 @@ import { Model } from "@anthropic-ai/sdk/resources/index.mjs"
 import { LARGE_MODEL_ID, SMALL_MODEL_ID, VISION_MODEL_ID } from "./constants"
 import { ModelConfiguration, ModelConfigurations } from "./host"
 import LLMS from "./llms.json"
-import { deleteEmptyValues } from "./util"
+import { deleteEmptyValues } from "./clone"
 
 export function defaultModelConfigurations(): ModelConfigurations {
     const aliases = [
