+++ packages/core/src/util.ts
@@ -63,21 +63,6 @@ export function deleteUndefinedValues<T extends Record<string, any>>(o: T): T {
     return o
 }
 
-export function deleteEmptyValues<T extends Record<string, any>>(o: T): T {
-    if (typeof o === "object")
-        for (const k in o) {
-            const v = o[k]
-            if (
-                v === undefined ||
-                v === null ||
-                v === "" ||
-                (Array.isArray(v) && !v.length)
-            )
-                delete o[k]
-        }
-    return o
-}
-
 export function collapseEmptyLines(text: string) {
     return text?.replace(/(\r?\n){2,}/g, "\n\n")
 }
